% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.r
\name{add2}
\alias{add2}
\title{An example addition function}
\usage{
add2(x1, x2)
}
\arguments{
\item{x1}{is the first number in the addition}

\item{x2}{is the second number in the addition}
}
\description{
An example addition function
}
