% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{birth_cohort_layouts}
\alias{birth_cohort_layouts}
\alias{data_birth_cohort_1983}
\title{Birth Cohort Layouts}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 119 rows and 7 columns.
}
\usage{
data_birth_cohort_1983
}
\description{
Layout data for birth cohort files.
}
\keyword{datasets}
