% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{births_layouts}
\alias{births_layouts}
\alias{data_births_1978}
\title{Births Layouts}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 127 rows and 7 columns.
}
\usage{
data_births_1978
}
\description{
Layout data for births files.
}
\keyword{datasets}
