% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cdc.r
\name{download_cdc}
\alias{download_cdc}
\title{download cdc data object}
\usage{
download_cdc(url)
}
\arguments{
\item{url}{is a cdc url path that can be found from \code{cdc_link_lookup} variable \code{us_data_url} path as \code{url} input.}
}
\value{
the temporary file where the file was downloaded.
}
\description{
Uses the \code{cdc_link_lookup} data frame.
}
\examples{
#'
\dontrun{
meta <- data_mortality_multiple_1969
download <- usdeaths::cdc_link_lookup
url <- download |>
  filter(section == "mortality_multiple" & year == 1969) |>
  pull(us_data_url)
path <- download_cdc(url)
}
}
