% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fetal_death_layouts}
\alias{fetal_death_layouts}
\alias{data_fetal_death_1982}
\title{Fetal Death Layouts}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 85 rows and 7 columns.
}
\usage{
data_fetal_death_1982
}
\description{
Layout data for fetal death files.
}
\keyword{datasets}
