% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_years.r
\name{get_available_years}
\alias{get_available_years}
\title{Get Available Years for a Data Type}
\usage{
get_available_years(
  type = c("mortality_multiple", "births", "birth_cohort", "fetal_deaths")
)
}
\arguments{
\item{type}{A character string specifying the data category.
Must be one of "mortality_multiple", "births", "birth_cohort", or "fetal_deaths".}
}
\value{
An integer vector of sorted years.
}
\description{
Queries the \code{usdeaths} package data index to identify years for which metadata
objects are available for a specific data category.
}
