% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_html_page.r
\name{get_html_page}
\alias{get_html_page}
\title{Get an HTML page from a URL}
\usage{
get_html_page(url)
}
\arguments{
\item{url}{Character string containing the URL of the web page to retrieve.}
}
\value{
An HTML document of class \code{xml_document}, ready for use with
\pkg{rvest} functions.
}
\description{
Downloads a web page using \pkg{httr2} and returns it as an HTML document.
This is useful for passing directly to scraping functions like \code{scrape_cdc_section()}.
}
\details{
The function performs a GET request using \pkg{httr2}, sets a user agent
to mimic a browser, and parses the response into an HTML document.
}
\examples{
\dontrun{
page <- get_html_page("https://www.cdc.gov/nchs/data_access/VitalStatsOnline.htm")
births <- scrape_cdc_section(
    page, "Births", "birth",
    c("User Guide", "U.S. Data", "U.S. Territories")
)
}

}
