% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdeaths-package.r
\docType{package}
\name{usdeaths-package}
\alias{usdeaths}
\alias{usdeaths-package}
\title{usdeaths: Scrape CDC Vital Statistics File Links}
\description{
Provides tools to scrape and parse downloadable file links from the CDC Vital Statistics Online data access pages, including births, mortality, fetal death, and related documentation.
}
\author{
\strong{Maintainer}: J. Hathaway \email{hathawayj@byui.edu} (\href{https://orcid.org/0000-0002-1574-0832}{ORCID})

Authors:
\itemize{
  \item Mitchell Mecham \email{mechamit000@gmail.com}
}

}
