% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{births_layouts}
\alias{births_layouts}
\alias{data_births_1978}
\alias{data_births_2018}
\alias{data_births_2019}
\alias{data_births_2020}
\alias{data_births_2021}
\alias{data_births_2022}
\alias{data_births_2023}
\alias{data_births_2024}
\title{Births Layouts}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 127 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 284 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 144 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 284 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 284 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 284 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 284 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 284 rows and 7 columns.
}
\usage{
data_births_1978

data_births_2018

data_births_2019

data_births_2020

data_births_2021

data_births_2022

data_births_2023

data_births_2024
}
\description{
Layout data for births files.
}
\keyword{datasets}
