% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{load_data}
\alias{load_data}
\title{read fixed width downloaded data}
\usage{
load_data(path, meta)
}
\arguments{
\item{path}{the file path returned from \code{download_cdc()} or the path to the zip file downloaded.}

\item{meta}{the \code{data_} object with the metadata needed to read the fixed width file format.}
}
\description{
read fixed width downloaded data
}
\details{
Uses output from \code{download_cdc()} to load the data into R as a tibble.
}
